Test that @TestScope works with parallel tests.
Checks that we support test runners which execute multiple tests
in parallel. Note that the current TestScope implementation only allows
use of the injector from the main test thread. This means individual
test cases which are multi-threaded will need to inject instances
up-front and then pass them down to other threads on their own. This
change does not alter this behaviour but hopefully the additional
test clarifies the intent of the existing code.